+++ trunk/src/RLE.cpp
@@ -168,7 +168,7 @@ void PartFileEncoderData::Decode(unsigned char *gapdata, int gaplen, unsigned ch
 	// in a first dword - real size
 	uint32 gapsize = ENDIAN_NTOHL( RawPeekUInt32( gapdata ) );
 	gapdata += sizeof(uint32);
-	m_gap_status.Realloc(gapsize*2*sizeof(uint32));
+	m_gap_status.Realloc(gapsize*2*sizeof(uint64));
 
 	m_gap_status.Decode(gapdata, gaplen - sizeof(uint32));
 }
